Added BoardView and updated ChessLogic
* Synced up both versions of chesslogic and added functionality for copys to be created.

* Added drawableID to ChessLogic.Piece

drawableID is used by BoardView to decide what to draw on each space. There's currently an ugly default texture for the rooks!
This won't compile until merged to master, as that's where all the PNG files are. Sorry.

* Added BoardView

Currently, it manually draws 64 plain black/white squares, but I'm open to using different colors or some cool textures.
Also, note that it WILL NOT COMPILE until ChessLogic is set up properly within the project package (the imports are all wrong).